DAT-1006 refactor cmdb objects
