+++ cmdb/docapi/docapi_template/docapi_template_base.py
@@ -15,11 +15,17 @@
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 """TODO: document"""
 import logging
+import json
+from pymongo import IndexModel
+
+from cmdb.database.utils import default
 # -------------------------------------------------------------------------------------------------------------------- #
 
 LOGGER = logging.getLogger(__name__)
 
-
+# -------------------------------------------------------------------------------------------------------------------- #
+#                                            TemplateManagementBase - CLASS                                            #
+# -------------------------------------------------------------------------------------------------------------------- #
 class TemplateManagementBase:
     """TODO: document"""
     ASCENDING = 1
@@ -35,6 +41,7 @@ class TemplateManagementBase:
 
     def __init__(self, **kwargs):
         self.public_id = None
+
         for key, value in kwargs.items():
             setattr(self, key, value)
 
@@ -42,10 +49,10 @@ class TemplateManagementBase:
     @classmethod
     def get_index_keys(cls):
         """TODO: document"""
-        from pymongo import IndexModel
         index_list = []
         for index in cls.INDEX_KEYS + cls.SUPER_INDEX_KEYS:
             index_list.append(IndexModel(**index))
+
         return index_list
 
 
@@ -55,8 +62,6 @@ class TemplateManagementBase:
         Returns:
             dict: json dump with database default encoding of the object attributes
         """
-        from cmdb.database.utils import default
-        import json
         return json.dumps(self.__dict__, default=default)
