+++ cmdb/framework/models/log.py
@@ -18,13 +18,13 @@ import logging
 from datetime import datetime
 from enum import Enum
 
-from cmdb.framework.cmdb_dao import CmdbDAO
+from cmdb.cmdb_objects.cmdb_dao import CmdbDAO
 from cmdb.framework.utils import Collection, Model
 # -------------------------------------------------------------------------------------------------------------------- #
 
 LOGGER = logging.getLogger(__name__)
 
-
+#TODO CLASS-FIX
 class LogAction(Enum):
     """TODO: document"""
     CREATE = 0
@@ -33,6 +33,7 @@ class LogAction(Enum):
     DELETE = 3
 
 
+#TODO CLASS-FIX
 class CmdbMetaLog(CmdbDAO):
     """CmdbMetaLog"""
     COLLECTION: Collection = 'framework.logs'
@@ -46,6 +47,7 @@ class CmdbMetaLog(CmdbDAO):
         super().__init__(public_id=public_id)
 
 
+#TODO CLASS-FIX
 class CmdbObjectLog(CmdbMetaLog):
     """TODO:document"""
 
@@ -154,6 +156,7 @@ class CmdbObjectLog(CmdbMetaLog):
         }
 
 
+#TODO CLASS-FIX
 class CmdbLog:
     """TODO: document"""
     REGISTERED_LOG_TYPE = {}
