+++ cmdb/manager/manager_provider.py
@@ -23,8 +23,8 @@ from cmdb.user_management.rights import __all__ as rights
 from cmdb.user_management.managers.right_manager import RightManager
 
 from cmdb.manager.categories_manager import CategoriesManager
-from cmdb.framework.cmdb_object_manager import CmdbObjectManager
-from cmdb.framework.managers.object_manager import ObjectManager
+from cmdb.manager.cmdb_object_manager import CmdbObjectManager
+from cmdb.manager.object_manager import ObjectManager
 from cmdb.manager.logs_manager import LogsManager
 from cmdb.docapi.docapi_template.docapi_template_manager import DocapiTemplateManager
 from cmdb.user_management.managers.user_manager import UserManager
@@ -32,7 +32,7 @@ from cmdb.user_management.managers.setting_manager import UserSettingsManager
 from cmdb.user_management.managers.group_manager import GroupManager
 from cmdb.media_library.media_file_manager import MediaFileManager
 from cmdb.exportd.exportd_logs.exportd_log_manager import ExportdLogManager
-from cmdb.framework.managers.type_manager import TypeManager
+from cmdb.manager.type_manager import TypeManager
 from cmdb.manager.locations_manager import LocationsManager
 from cmdb.manager.section_templates_manager import SectionTemplatesManager
 from cmdb.manager.object_links_manager import ObjectLinksManager
